update block id from listening modals
current, the serial order in which a user reads the documentation sets the
opening_modals example as a pre-requisite that the listening_modals builds
upon.

Currently, an attempt to walk through the examples in order yeilds a
python

`keyError: 'block_c'`

This is because block_c does not match the name of the previous modal examle
`input_c`.

This patch updates the variable to be the same so that users can step through
the documentation without the above exception.